% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_target.R
\name{align_target}
\alias{align_target}
\title{Align microbiome reads to a set of reference libraries}
\usage{
align_target(reads, libs,
  project_name = tools::file_path_sans_ext(reads), threads = 8,
  mismatch = 5)
}
\arguments{
\item{reads}{Location to the .fastq file to align}

\item{libs}{A list of Subread index headers for alignment}

\item{project_name}{A name for the project, which names the output .bam file (e.g. project_name.bam). Defaults to the basename of the reads file.}

\item{threads}{The number of threads for the Subread alignment. Defaults to 8}

\item{mismatch}{Numeric value giving the maximum number of mis-matched bases allowed in the alignment. Default is 3. Mis-matches found in soft-clipped bases are not counted.}
}
\value{
This function writes to file a merged and sorted .bam file after aligning to all reference libraries given. The function also outputs the new .bam filename.
}
\description{
This is the main animalcules target library mapping function, using Rsubread and multiple libraries. Aligns to each library separately, filters unmapped reads from each file, and then merges and sorts the .bam files from each library into one output file.
}
\examples{
## Get a reference genome library
download_refseq('viral', compress = FALSE)

## Make and align to a single a reference genome library
mk_subread_index('viral.fasta')
readPath <- system.file("extdata", "virus_example.fastq", package = "animalcules")
viral_map <- align_target( readPath, "viral", "virus_example")
viral_map_sam <- Rsamtools::asSam(viral_map, overwrite=TRUE)

## Make and align to a multiple reference genome libraries
mk_subread_index('viral.fasta', split=0.0005)
targLibs <- c("viral_1", "viral_2")
readPath <- system.file("extdata", "virus_example.fastq", package = "animalcules")
viral_map <- align_target( readPath, targLibs, "virus_example")

}
