% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_calc_cors.R
\name{calc_cors}
\alias{calc_cors}
\title{Calculates Spearman rank-based correlation and corresponding p values}
\usage{
calc_cors(df1, df2, no.samples)
}
\arguments{
\item{df1}{Dataframe containing sample data, where each sample has its own row}

\item{df2}{Second dataframe, against which df1 will be correlated}

\item{no.samples}{Number of samples being considered}
}
\value{
A list containing the correlation matrix and its associated t-values
}
\description{
Calculates Spearman rank-based correlation and corresponding p values
}
\examples{
x <- matrix(rnorm(200), ncol = 10, nrow = 20)
y <- matrix(rnorm(200), ncol = 10, nrow = 20)
results <- calc_cors(x, y, no.samples = 20)

results$cors
results$ts

}
