% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_calc_sig.R
\name{calc_sig}
\alias{calc_sig}
\title{Calculate significant correlations, given hostExpression data and microbial abundance data}
\usage{
calc_sig(cors, ps, adjust, no.sig = 1, alpha = 0.05)
}
\arguments{
\item{cors}{A matrix of correlation coefficients}

\item{ps}{A matrix of the p-values for each correlation}

\item{adjust}{p-value correction method. Select one from `p.adjust.methods`}

\item{no.sig}{Minimum number of significant correlations to be extracted. Default is set to 0.05}

\item{alpha}{Threshold for significance. Default is set to 0.05}
}
\value{
Subset of cors, containing only the correlation coefficients that are significant
}
\description{
Calculate significant correlations, given hostExpression data and microbial abundance data
}
\examples{
x <- matrix(rnorm(200), ncol = 10, nrow = 20)
y <- matrix(rnorm(200), ncol = 10, nrow = 20)
no.samples <- 20
total_correlations <- calc_cors(x, y, no.samples = 20)
cors <- total_correlations[[1]]
ps <- total_correlations[[2]]
adjust <- 'bonferroni'
no.sig <- 3
alpha <- 0.05
sig_correlations <- calc_sig(cors,
                             ps,
                             adjust,
                             no.sig,
                             alpha
)


}
