% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_corr_func.R
\name{corr_func}
\alias{corr_func}
\title{Calculate significant correlations between microbial abundance and another assay}
\usage{
corr_func(
  MAE,
  asys,
  tax_level = NA,
  no.sig = 1,
  correction = "bonferroni",
  hide_ax = NA
)
}
\arguments{
\item{MAE}{A MultiAssayExperiment object}

\item{asys}{List of assays to calculate correlations for. Minimum length = 1}

\item{tax_level}{Taxonomic level of interest. Maximum length = 2, default is set to "genus"}

\item{no.sig}{Minimum number of significant correlations to be extracted}

\item{correction}{Method for p-value correction for multiple hypotheses. Default is bonferroni}

\item{alpha}{If no correction is specified, will use alpha to determine significance}
}
\value{
a list with plotly heatmap, summary tables, and correlation matrix
}
\description{
Calculate significant correlations between microbial abundance and another assay
}
\examples{
library(SummarizedExperiment)
data_dir = system.file('extdata/MAE.rds', package = 'animalcules')
toy_data <- readRDS(data_dir)
results <- corr_func(MAE = toy_data, asys = c("MicrobeGenetics", "hostExpression"))
results$plot # heatmap of significant correlations
result$summary # summary of the significantly correlated groups
results$summary_t10 # top 10 groups (by size)
results$cormat # correlation matrix

}
