% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{counts_to_relabu}
\alias{counts_to_relabu}
\title{Covert a counts table to a relative abundances table}
\usage{
counts_to_relabu(counts_table)
}
\arguments{
\item{counts_table}{A organism x sample data frame of counts}
}
\value{
A organism x sample data frame of relative abundances
}
\description{
Covert a counts table to a relative abundances table
}
\examples{
toy_data <- readRDS("data/MAE.rds")
microbe <- toy_data[['MicrobeGenetics']] #double bracket subsetting is easier
counts_table <- as.data.frame(assays(microbe))[,rownames(sam_table)] # organism x sample
relabu <- counts_to_relabu(counts_table)

}
