% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_abundance.R
\name{differential_abundance}
\alias{differential_abundance}
\title{Differential abundance analysis}
\usage{
differential_abundance(MAE, tax_level, input_da_condition = c(),
  input_da_condition_covariate = NULL, min_num_filter = 5,
  input_da_padj_cutoff = 0.05)
}
\arguments{
\item{MAE}{A multi-assay experiment object}

\item{tax_level}{The taxon level used for organisms}

\item{input_da_condition}{Which condition is the target condition}

\item{input_da_condition_covariate}{Covariates added to linear function}

\item{min_num_filter}{Minimum number of counts of reads mapped to this microbe}

\item{input_da_padj_cutoff}{adjusted pvalue cutoff}
}
\value{
A plotly object
}
\description{
Differential abundance analysis
}
\examples{
toy_dat_path <- system.file("data", "MAE.rds", package = "animalcules")
toy_data <- readRDS(toy_dat_path)
p <- differential_abundance(toy_data,
tax_level="genus",
input_da_condition=c("DISEASE"),
input_da_condition_covariate=c("SEX"),
min_num_filter = 5,
input_da_padj_cutoff = 0.05)
p

}
