% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred_umap.R
\name{dimred_umap}
\alias{dimred_umap}
\title{Dimensionality reduction through UMAP}
\usage{
dimred_umap(MAE, tax_level, color, shape = NULL, k = c("2D", "3D"),
  n_neighbors = 15, min_dist = 3, datatype = c("counts", "relabu",
  "logcpm"))
}
\arguments{
\item{MAE}{A multi-assay experiment object}

\item{tax_level}{The taxon level used for organisms}

\item{color}{A condition to color data points by e.g. "AGE"}

\item{shape}{A condition to shape data points by e.g. "SEX"}

\item{k}{Plot dimensions e.g. c("2D","3D")}

\item{n_neighbors}{The number of nearest neighbors}

\item{min_dist}{Determines how close points appear in the final layout}

\item{datatype}{Datatype to use e.g. c("counts", "relabu", "logcpm")}
}
\value{
A plotly object
}
\description{
Dimensionality reduction through UMAP
}
\examples{
data_dir = system.file("extdata/MAE.rds", package = "animalcules")
toy_data <- readRDS(data_dir)
p <- dimred_umap(toy_data,
                 tax_level="phylum",
                 color="AGE",
                 shape="GROUP",
                 k="3D",
                 n_neighbors=15,
                 min_dist=3,
                 datatype="logcpm")
p

}
