% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimredg_pca.R
\name{dimredg_pca}
\alias{dimredg_pca}
\title{Dimensionality reduction through PCA, for gene expression assays}
\usage{
dimredg_pca(
  MAE,
  color,
  shape = NULL,
  pcx = 1,
  pcy = 2,
  pcz = NULL,
  datatype = c("logcpm", "counts")
)
}
\arguments{
\item{MAE}{A multi-assay experiment object}

\item{color}{A condition to color data points by e.g. "AGE"}

\item{shape}{A condition to shape data points by e.g. "SEX"}

\item{pcx}{Principal component on the x-axis e.g. 1}

\item{pcy}{Principal component on the y-axis e.g. 2}

\item{pcz}{Principal component on the z-axis e.g. 3}

\item{datatype}{Datatype to use e.g. c("logcpm", "counts")}
}
\value{
A list with a plotly object and summary table
}
\description{
Dimensionality reduction through PCA, for gene expression assays
}
\examples{
data_dir = system.file("extdata/MAE.rds", package = "animalcules")
toy_data <- readRDS(data_dir)
result <- dimredg_pca(toy_data,
                      color="AGE",
                      shape="DISEASE",
                      pcx=1,
                      pcy=2,
                      datatype="logcpm")
result$plot
result$table

}
