% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_beta_NMDS.R
\name{diversity_beta_NMDS}
\alias{diversity_beta_NMDS}
\title{Beta diversity NMDS plot}
\usage{
diversity_beta_NMDS(
  MAE,
  tax_level,
  input_beta_method,
  input_select_beta_condition
)
}
\arguments{
\item{MAE}{A multi-assay experiment object}

\item{tax_level}{The taxon level used for organisms}

\item{input_beta_method}{bray, jaccard}

\item{input_select_beta_condition}{Which condition to group samples}
}
\value{
A plotly object/NMDS plot
}
\description{
Beta diversity NMDS plot
}
