% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_refseq.R
\name{download_kingdom_summary}
\alias{download_kingdom_summary}
\title{Download RefSeq assembly summary}
\usage{
download_kingdom_summary()
}
\arguments{
\item{kingdom}{Select the kingdom taxonomy to download. Options are 'archaea', 'bacteria', 'fungi', 'invertebrate', 'plant', 'protozoa', 'vertibrate', 'vertibrate_other', 'viral'}
}
\value{
Returns (eventually) a list of possible subgroups for download
}
\description{
This function will automatically download RefSeq assembly file for a specified kindom. This the precursor for the function that will allow the user to select subgroups of the kingdom for download.
}
\examples{
## Download all RefSeq reference bacterial genomes
download_kingdom_summary('bacteria')

}
