% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_summary_top.R
\name{filter_summary_top}
\alias{filter_summary_top}
\title{Data visualization by pie chart}
\usage{
filter_summary_top(MAE, samples_discard, filter_type, sample_condition)
}
\arguments{
\item{MAE}{A multi-assay experiment object}

\item{samples_discard}{The list of samples to filter}

\item{filter_type}{Either 'By Microbes' or 'By Metadata'}

\item{sample_condition}{Which condition to check e.g. 'AGE}
}
\value{
A plotly object
}
\description{
Data visualization by pie chart
}
\examples{
data_dir = system.file('extdata/MAE.rds', package = 'animalcules')
toy_data <- readRDS(data_dir)
result <- filter_summary_top(toy_data,
                             samples_discard = c('subject_2', 'subject_4'),
                             filter_type = 'By Microbes',
                             sample_condition = 'AGE')
result

}
