% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_taxonomy_300.R
\name{find_taxonomy_300}
\alias{find_taxonomy_300}
\title{Find the taxonomy for maximum 300 tids}
\usage{
find_taxonomy_300(tids)
}
\arguments{
\item{tids}{Given taxonomy ids}
}
\value{
taxondata Data with the taxonomy information
}
\description{
Find the taxonomy for maximum 300 tids
}
\examples{
\donttest{
taxonLevels <- find_taxonomy_300(tids = 1200)
}

}
