% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes_dimred_gsva.R
\name{genes_dimred_gsva}
\alias{genes_dimred_gsva}
\title{Produces GSVA scores using gene sets generated from user-selected MSigDB pathways}
\usage{
genes_dimred_gsva(MAE, collection, sub_collection = NA)
}
\arguments{
\item{MAE}{MultiAssayExperiment containing a "hostExpression" SummarizedExperiment object}

\item{collection}{MSigDB collection of gene sets}

\item{sub_collection}{MSigDB sub-collection. Only available for some collections}
}
\value{
A SummarizedExperiment object containing the ssGSEA scores of hostExpression
}
\description{
Produces GSVA scores using gene sets generated from user-selected MSigDB pathways
}
