% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_heatmap_cors.R
\name{heatmap_cors}
\alias{heatmap_cors}
\title{Create a plotly heatmap of correlations}
\usage{
heatmap_cors(cormat, hide_ax)
}
\arguments{
\item{cormat}{A matrix of significant correlations}

\item{hide_ax}{Option to hide axes. NA, 'xax' (x-axis), 'yax' (y-axis), or 'bax' (both axes)}
}
\value{
A plotly heatmap of the signficant correlations
}
\description{
Create a plotly heatmap of correlations
}
\examples{
x <- matrix(rnorm(200), ncol = 10, nrow = 20)
y <- matrix(rnorm(200), ncol = 10, nrow = 20)
no.samples <- 20
total_correlations <- calc_cors(x, y, no.samples = 20)
cors <- total_correlations$cors
ts <- total_correlations$ts
alpha <- 0.05
t_crit <- abs(stats::qt(alpha, no.samples-2))
sig_correlations <- calc_sig(cors, ts, no.sig, t_crit)

p <- heatmap_cors(sig_correlations)
p

}
