% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_proj_insert_into_MAE.R
\name{insert_into_MAE}
\alias{insert_into_MAE}
\title{Inserting GSVA scores into a MultiAssayExperiment}
\usage{
insert_into_MAE(MAE, gsva_data)
}
\arguments{
\item{MAE}{MultiAssayExperiment object}

\item{gsva_data}{ssGSEA scores calculated from the path_proj function}
}
\value{
Updated MAE containing the ssGSEA scores for the user-selected geneset
}
\description{
Inserting GSVA scores into a MultiAssayExperiment
}
\examples{
library(SummarizedExperiment)
data_dir = system.file('extdata/MAE.rds', package = 'animalcules')
toy_data <- readRDS(data_dir)
gsva_results <- path_proj(toy_data, "H", NA) # H = Hallmark gene sets

test_MAE <- insert_into_MAE(toy_data, gsva_results)

}
