% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.categorical}
\alias{is.categorical}
\title{Check if object is categorical}
\usage{
is.categorical(v)
}
\arguments{
\item{x}{A single value}
}
\value{
Boolean
}
\description{
Check if object is categorical
}
\examples{
nums <- 2
is.categorical(nums)

}
