% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.integer1}
\alias{is.integer1}
\title{check if integer(1)}
\usage{
is.integer1(x)
}
\arguments{
\item{x}{A single value}
}
\value{
Boolean
}
\description{
check if integer(1)
}
\examples{
nums <- 2
is.integer1(nums)

}
