% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{MAE}
\alias{MAE}
\title{Toy Host/Microbiome MultiAssayExperiment (MAE)}
\format{MAE object with 4 data attributes
\describe{
  \item{ExperimentList}{A list of experiments represented as SummarizedExperiment objects}
  \item{colData}{A dataframe of samples and phenotypic attributes}
  \item{sampleMap}{A dataframe of samples mapped to experiments}
  \item{metadata}{Metadata information regarding experiments}
}}
\usage{
MAE(pred, obs, na.rm = FALSE)
}
\description{
A toy multi-assay experiment object with mock data
that is useful for tesing animacules functionality
}
