% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_subread_index.R
\name{mk_subread_index}
\alias{mk_subread_index}
\title{Make a Subread index}
\usage{
mk_subread_index(ref_lib, split = 4, mem = 8000)
}
\arguments{
\item{ref_lib}{The name/location of the reference library file, in (uncompressed) .fasta format}

\item{split}{The maximum allowed size of the genome file (in GB). If the ref_lib file is larger than this, the function will spolt the library into multiple parts}

\item{mem}{The maximum amount of memory (in MB) that can be used by the index generation process (used by the Rsubread::buildindex function)}
}
\value{
Returns one or more Subread indexes for the supplied reference .fasta file. If multiple indexes are created, the libraries will be named the ref_lib basename plus _1, _2, etc.
}
\description{
This function is a wrapper for the Rsubread::buildindex function. It will generate one or more Subread indexes from a .fasta file. If the library is too large (default >4GB) it will automatically be split into multiple indexes, with _1, _2, etc at the end of the ref_lib basename.
}
\examples{
## Download all RefSeq reference viral genomes and make an index
download_refseq('viral', compress = FALSE)
mk_subread_index('viral.fasta')

## Download all RefSeq reference viral genomes and make more than one index
download_refseq('viral', compress = FALSE)
mk_subread_index('viral.fasta', split = .0005)

}
