% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_proj.R
\name{path_proj}
\alias{path_proj}
\title{Pathway projection using GSVA}
\usage{
path_proj(MAE, collection, sub_collection)
}
\arguments{
\item{MAE}{MultiAssayExperiment having at least 1 gene expression assay (containing gene symbols)}

\item{collection}{User-specified MSigDB collection}

\item{sub_collection}{Optional, user-specified MSigDB sub-collection}
}
\value{
SummarizedExperiment containing ssGSEA scores of selected gene sets
}
\description{
Pathway projection using GSVA
}
\examples{
library(SummarizedExperiment)
data_dir = system.file('extdata/MAE.rds', package = 'animalcules')
toy_data <- readRDS(data_dir)

gsva_results <- path_proj(toy_data, "H", NA) # H = Hallmark gene sets

}
