% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pct2str}
\alias{pct2str}
\title{Converts decimal percentage to string with specified digits}
\usage{
pct2str(v, digits = 2)
}
\arguments{
\item{x}{A single value}
}
\value{
Boolean
}
\description{
Converts decimal percentage to string with specified digits
}
