% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{percent}
\alias{percent}
\title{Format decimals to percentages}
\usage{
percent(x, digits = 2, format = "f")
}
\arguments{
\item{x}{An array of decimals}
}
\value{
An array of formatted strings
}
\description{
Format decimals to percentages
}
