% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_rd_MAE.R
\name{rd_MAE}
\alias{rd_MAE}
\title{Reading in a MultiAssayExperiment object}
\usage{
rd_MAE(MAE, asys)
}
\arguments{
\item{MAE}{A MultiAssayExperiment object}

\item{asys}{Character vector of assays to be selected for correlation analysis}
}
\value{
A list containing the subsetted MultiAssayExperiment object, as well as the number of samples found in all assays specified
}
\description{
Reading in a MultiAssayExperiment object
}
\examples{
library(SummarizedExperiment)
data_dir = system.file('extdata/MAE.rds', package = 'animalcules')
toy_data <- readRDS(data_dir)
results <- rd_MAE(toy_data, c("MicrobeGenetics", "HostGenetics"))

results[[1]] # subsetted MultiAssayExperiment
results[[2]] # total number of samples

}
