% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_separate_by_condition.R
\name{separate_by_condition}
\alias{separate_by_condition}
\title{Splits a MultiAssayExperiment object according to its colData}
\usage{
separate_by_condition(MAE, condition)
}
\arguments{
\item{MAE}{A MultiAssayExperiment object}

\item{condition}{One of the covariates listed in the colData of the MultiAssayExperiment. Must have two different levels.}
}
\value{
Two MultiAssayExperiment objects, split apart from the original
}
\description{
Splits a MultiAssayExperiment object according to its colData
}
\examples{
library(MultiAssayExperiment)
data_dir = system.file('extdata/MAE.rds', package = 'animalcules')
toy_data <- readRDS(data_dir)
colData(toy_data) # "DISEASE" is a condition with 2 levels
toy_data_split <- separate_by_condition(toy_data, "DISEASE")
toy_data_split[1]
toy_data_split[2]

}
