% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_subset_cormat_by_row.R
\name{subset_cormat_by_row}
\alias{subset_cormat_by_row}
\title{Subsetting a correlation matrix by user-selected rows}
\usage{
subset_cormat_by_row(rows_selected, summary, cormat)
}
\arguments{
\item{rows_selected}{Row indices selected by the user}

\item{summary}{Summary table produced by corr_func}

\item{cormat}{Correlation matrix produced by corr_func}
}
\value{
A subsetted correlation matrix containing only the rows specified by the user
}
\description{
Subsetting a correlation matrix by user-selected rows
}
\examples{
library(SummarizedExperiment)
data_dir = system.file('extdata/MAE.rds', package = 'animalcules')
toy_data <- readRDS(data_dir)
results <- corr_func(MAE = toy_data, 
                     asys = c('MicrobeGenetics', 'hostExpression'),
                     tax_level='genus',
                     correction='BH', # Benjamini-Hochberg
                     alpha=0.1,
                     no.sig=1 # only want microbes correlated with =>20 genes
                     )
rows_selected <- c(1, 2, 3)

sub_cormat <- subset_cormat_by_row(rows_selected, results$summary, results$cormat)
}
