% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_summary_cors.R
\name{summary_cors}
\alias{summary_cors}
\title{Summarize significant correlations}
\usage{
summary_cors(cormat)
}
\arguments{
\item{cormat}{A matrix of significant correlations}
}
\value{
A summary table of the correlated groups, as well as how big each group is
}
\description{
Summarize significant correlations
}
