% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_biom.R
\name{write_to_biom}
\alias{write_to_biom}
\title{Output biom}
\usage{
write_to_biom(MAE, path_to_output)
}
\arguments{
\item{MAE}{A multi-assay experiment object}

\item{path_to_output}{The folder to output biom file}
}
\value{
A message
}
\description{
Output biom
}
\examples{
data_dir = system.file('extdata/MAE.rds', package = 'animalcules')
toy_data <- readRDS(data_dir)
write_to_biom(toy_data, ".")


}
