% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalcules-package.R
\docType{package}
\name{animalcules-package}
\alias{animalcules}
\alias{animalcules-package}
\title{animalcules: Statistical microbiome analysis on metagenomics data}
\description{
Statistical microbiome analysis on metagenomics data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/compbiomed/animalcules}
  \item Report bugs at \url{https://github.com/compbiomed/animalcules/issues}
}

}
\keyword{internal}
