% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalcules-package.R
\docType{package}
\name{animalcules-package}
\alias{animalcules}
\alias{animalcules-package}
\title{animalcules: What the Package Does (one line, title case)}
\description{
Lorem ipsum dolor sit amet, consectetur adipiscing elit. Etiam vitae sapien
    at odio vehicula hendrerit. Curabitur finibus scelerisque ex volutpat
    rhoncus. Mauris vel arcu placerat dui posuere mattis. Mauris laoreet pretium
    purus, eu volutpat tellus egestas vel. Quisque ornare eros ut nulla finibus
    lobortis. Praesent dictum pellentesque est a mattis. Suspendisse id rhoncus
    purus.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/schifferl/animalcules}
  \item Report bugs at \url{https://github.com/schifferl/animalcules/issues}
}

}
\keyword{internal}
