% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_calc_sig.R
\name{calc_sig}
\alias{calc_sig}
\title{Calculate significant correlations, given hostExpression data and microbial abundance data}
\usage{
calc_sig(cors, ts, no.sig, t_crit)
}
\arguments{
\item{cors}{A matrix of correlation coefficients}

\item{ts}{A matrix of t-values, associated with each correlation coefficient in cors}

\item{no.sig}{Minimum number of significant correlations to be extracted}

\item{t_crit}{Critical t-value for determining significance}
}
\value{
Subset of cors, containing only the correlation coefficients that are significant
}
\description{
Calculate significant correlations, given hostExpression data and microbial abundance data
}
\examples{
x <- matrix(rnorm(200), ncol = 10, nrow = 20)
y <- matrix(rnorm(200), ncol = 10, nrow = 20)
no.samples <- 20
total_correlations <- calc_cors(x, y, no.samples = 20)
cors <- total_correlations$cors
ts <- total_correlations$ts
alpha <- 0.05
t_crit <- abs(stats::qt(alpha, no.samples-2))
sig_correlations <- calc_sig(cors, ts, no.sig, t_crit)

}
