% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_corr_func.R
\name{corr_func}
\alias{corr_func}
\title{Calculate significant correlations between microbial abundance and another assay}
\usage{
corr_func(
  MAE,
  asys,
  tax_level = "genus",
  no.sig = 1,
  correction = "bonferroni",
  alpha = 0
)
}
\arguments{
\item{MAE}{A MultiAssayExperiment object}

\item{asys}{List of assays to calculate correlations for. Minimum length = 1}

\item{tax_level}{Taxonomic level of interest. Maximum length = 2, default is set to "genus"}

\item{no.sig}{Minimum number of significant correlations to be extracted}

\item{correction}{Method for p-value correction for multiple hypotheses. Default is bonferroni}

\item{alpha}{If no correction is specified, will use alpha to determine significance}
}
\description{
Calculate significant correlations between microbial abundance and another assay
}
\examples{
library(SummarizedExperiment)
data_dir = system.file('extdata/MAE.rds', package = 'animalcules')
toy_data <- readRDS(data_dir)
correlation_matrix <- corr_func(MAE = toy_data, asys = c(MicrobeGenetics, HostExpression), tax_level = "genus")

}
