% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_corr_func.R
\name{corr_func}
\alias{corr_func}
\title{Calculate significant correlations two assays}
\usage{
corr_func(
  MAE,
  asys,
  tax_level = NA,
  no.sig = 1,
  correction = "bonferroni",
  alpha = 0.05
)
}
\arguments{
\item{MAE}{A MultiAssayExperiment object}

\item{asys}{List of assays to calculate correlations for. Minimum length = 1}

\item{tax_level}{Taxonomic level of interest. Maximum length = 2.}

\item{no.sig}{Minimum number of significant correlations}

\item{correction}{Method for p-value correction for multiple hypotheses, selecting from one of `stats::p.value.methods`. Default is 'bonferroni'.}

\item{alpha}{Significance threshold. Default is 0.05}
}
\value{
a list with summary tables and correlation matrix
}
\description{
Calculate significant correlations two assays
}
\examples{
library(SummarizedExperiment)
data_dir = system.file('extdata/MAE.rds', package = 'animalcules')
toy_data <- readRDS(data_dir)
results <- corr_func(MAE = toy_data, 
                     asys = c('MicrobeGenetics', 'hostExpression'),
                     tax_level='genus',
                     correction='BH', # Benjamini-Hochberg
                     alpha=0.1,
                     no.sig=1 # only want microbes correlated with =>20 genes
                     )
results$summary # summary of the significantly correlated groups
results$cormat # correlation matrix

}
