% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_corr_network.R
\name{corr_network}
\alias{corr_network}
\title{Create a co-occurence network, given a counts table and correlation matrix}
\usage{
corr_network(MAE, assay, cormat, group)
}
\arguments{
\item{MAE}{A MultiAssayExperiment object}

\item{assay}{SummarizedExperiment of interest, within the MultiAssayExperiment}

\item{cormat}{Correlation matrix produced by corr_func}

\item{group}{Correlation group to produce network for}
}
\value{
qgraph network of the co-occurences within the group selected
}
\description{
Create a co-occurence network, given a counts table and correlation matrix
}
\examples{
data_dir = system.file('extdata/MAE.rds', package = 'animalcules')
toy_data <- readRDS(data_dir)
results <- corr_func(toy_data, 
                     asys = c('MicrobeGenetics', 'hostExpression'),
                     tax_level="genus")
group <- 'Actinomyces' # microbe from results$summary
fig <- corr_network(MAE = toy_data, 
                    assay = "hostExpression", 
                    cormat = results$cormat, 
                    group = group)
fig

}
