% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_enrich_cors.R
\name{enrich_cors}
\alias{enrich_cors}
\title{Run enrichR analysis on significantly correlated groups using hypeR}
\usage{
enrich_cors(corr_results, group_selected, geneset_db)
}
\arguments{
\item{corr_results}{Results from corr_func. Second assay selected must contain gene expression values}

\item{group_selected}{User input, which group to run in Enrichr}

\item{geneset_db}{User input, which database to use for Enrichr analysis}
}
\value{
Plotly displaying enrichment terms and scores
}
\description{
Run enrichR analysis on significantly correlated groups using hypeR
}
\examples{
data_dir = system.file('extdata/MAE.rds', package = 'animalcules')
toy_data <- readRDS(data_dir)
results <- corr_func(toy_data, 
                     asys = c('MicrobeGenetics', 'HostGenetics'),
                     tax_level="genus")
                     
group <- 'Acinetobacter' # microbe from results

db <- 'KEGG_2019_Human' # Enrichr database

p <- enrich_cors(results$summary, group, db) # Run enrichment
p

}
