% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_taxonomy.R
\name{find_taxonomy}
\alias{find_taxonomy}
\title{Find the taxonomy for unlimited tids}
\usage{
find_taxonomy(tids)
}
\arguments{
\item{tids}{Given taxonomy ids}
}
\value{
taxondata Data with the taxonomy information
}
\description{
Find the taxonomy for unlimited tids
}
\examples{
taxonLevels <- find_taxonomy(1200)
}
