% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_taxonomy_300.R
\name{find_taxonomy_300}
\alias{find_taxonomy_300}
\title{Find the taxonomy for maximum 300 tids}
\usage{
find_taxonomy_300(tids)
}
\arguments{
\item{tids}{Given taxonomy ids}
}
\value{
taxondata Data with the taxonomy information
}
\description{
Find the taxonomy for maximum 300 tids
}
\examples{
example_data_dir <- system.file("example/data", package = "PathoStat")
pathoreport_file_suffix <- "-sam-report.tsv"
datlist <- read_pathoscope_data(example_data_dir,
pathoreport_file_suffix, input.files.name.vec = as.character(1:6))
dat <- datlist$data
ids <- rownames(dat)
tids <- unlist(lapply(ids, FUN = grepTid))
taxonLevels <- find_taxonomy_300(tids[1:5])
}
