% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_heatmap_cors.R
\name{heatmap_cors}
\alias{heatmap_cors}
\title{Plot heatmap given a correlation matrix}
\usage{
heatmap_cors(cormat, hide_ax)
}
\arguments{
\item{cormat}{Correlation matrix with significant correlations}

\item{hide_ax}{Which axis to hide}
}
\value{
Clustered heatmap, visualizing the correlations
}
\description{
Plot heatmap given a correlation matrix
}
\examples{
library(SummarizedExperiment)
data_dir = system.file('extdata/MAE.rds', package = 'animalcules')
toy_data <- readRDS(data_dir)
results <- corr_func(MAE = toy_data, 
                     asys = c('MicrobeGenetics', 'HostGenetics'),
                     tax_level='genus',
                     correction='BH', # Benjamini-Hochberg
                     alpha=0.1,
                     no.sig=1 # only want microbes correlated with =>20 genes
                     )
p <- heatmap_cors(results$cormat, hide_ax=NA)
p

}
