% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAE.R
\docType{data}
\name{MAE}
\alias{MAE}
\title{Host/Microbiome Toy MultiAssayExperiment}
\format{MultiAssayExperiment}
\usage{
MAE(pred, obs, na.rm = FALSE)
}
\description{
The MultiAssayExperiment object
}
\examples{
data("MAE")
}
